backend-api/requirements.txt
Django
django-ninja
rav
django-ninja-jwt[crypto]
pydantic[email]
django-cors-headers
python-decouple